```
Lightning CSS:    5 ms   (7.5 times faster)
Stylis:           6 ms   (6.3 times faster)
Autoprefixer dev: 29 ms  (1.2 times faster)
Autoprefixer:     36 ms
Stylecow:         204 ms (5.7 times slower)
```

- do cheap checks before expensive checks
- loop inside walkers instead of the other way around